augmentation的验证与使用
