--- augtools/img/run.py
@@ -1,1 +1,0 @@
-print(16/3)
