+++ augtools/img/transforms/blur/JPEG.py
@@ -1,15 +1,15 @@
 from augtools.img.transform import ImageTransform
 from augtools.img.transforms.utils.img_utils import *
-import skimage as sk
 from io import BytesIO
 from PIL import Image as PILImage
+
 
 class JPEG(ImageTransform):
     def __init__(
         self,
         always_apply: bool = False,
         p: float = 0.5,
-        severity = 1,
+        severity: int = 1,
     ):
         super().__init__(always_apply=always_apply, p=p)
         self.severity = severity
