+++ augtools/img/transforms/blur/glass_blur_transform.py
@@ -1,18 +1,14 @@
 from augtools.img.transform import ImageTransform
 from augtools.img.transforms.utils.img_utils import *
-import skimage
 from skimage.filters import gaussian
 
 
 class GlassBlur(ImageTransform):
-    """
-
-    """
     def __init__(
         self,
         always_apply: bool = False,
         p: float = 0.5,
-        severity = 1,
+        severity: int = 1,
     ):
         super().__init__(always_apply=always_apply, p=p)
         self.severity = severity
