+++ augtools/img/transforms/geometric/ElasticTransform.py
@@ -1,231 +1,8 @@
 from augtools.img.transform import DualTransform
 from augtools.img.transforms.utils.img_utils import *
-import skimage as sk
 import random
-from typing_extensions import Concatenate, ParamSpec
-import math
-import random_utils
-from scipy.ndimage import gaussian_filter
-
-
-ImageColorType = Union[float, Sequence[float]]
-
-
-def get_num_channels(image: np.ndarray) -> int:
-    return image.shape[2] if len(image.shape) == 3 else 1
-
-
-def _maybe_process_in_chunks(
-    process_fn, **kwargs
-) -> Callable[[np.ndarray], np.ndarray]:
-    """
-    Wrap OpenCV function to enable processing images with more than 4 channels.
-
-    Limitations:
-        This wrapper requires image to be the first argument and rest must be sent via named arguments.
-
-    Args:
-        process_fn: Transform function (e.g cv2.resize).
-        kwargs: Additional parameters.
-
-    Returns:
-        numpy.ndarray: Transformed image.
-
-    """
-
-    @wraps(process_fn)
-    def __process_fn(img: np.ndarray) -> np.ndarray:
-        num_channels = get_num_channels(img)
-        if num_channels > 4:
-            chunks = []
-            for index in range(0, num_channels, 4):
-                if num_channels - index == 2:
-                    # Many OpenCV functions cannot work with 2-channel images
-                    for i in range(2):
-                        chunk = img[:, :, index + i : index + i + 1]
-                        chunk = process_fn(chunk, **kwargs)
-                        chunk = np.expand_dims(chunk, -1)
-                        chunks.append(chunk)
-                else:
-                    chunk = img[:, :, index : index + 4]
-                    chunk = process_fn(chunk, **kwargs)
-                    chunks.append(chunk)
-            img = np.dstack(chunks)
-        else:
-            img = process_fn(img, **kwargs)
-        return img
-
-    return __process_fn
-
-
-def elastic_transform(
-    img: np.ndarray,
-    alpha: float,
-    sigma: float,
-    alpha_affine: float,
-    interpolation: int = cv2.INTER_LINEAR,
-    border_mode: int = cv2.BORDER_REFLECT_101,
-    value: Optional[ImageColorType] = None,
-    random_state: Optional[np.random.RandomState] = None,
-    approximate: bool = False,
-    same_dxdy: bool = False,
-):
-    """Elastic deformation of images as described in [Simard2003]_ (with modifications).
-    Based on https://gist.github.com/ernestum/601cdf56d2b424757de5
-
-    .. [Simard2003] Simard, Steinkraus and Platt, "Best Practices for
-         Convolutional Neural Networks applied to Visual Document Analysis", in
-         Proc. of the International Conference on Document Analysis and
-         Recognition, 2003.
-    """
-    height, width = img.shape[:2]
-
-    # Random affine
-    center_square = np.array((height, width), dtype=np.float32) // 2
-    square_size = min((height, width)) // 3
-    alpha = float(alpha)
-    sigma = float(sigma)
-    alpha_affine = float(alpha_affine)
-
-    pts1 = np.array(
-        [
-            center_square + square_size,
-            [center_square[0] + square_size, center_square[1] - square_size],
-            center_square - square_size,
-        ],
-        dtype=np.float32,
-    )
-    pts2 = pts1 + random_utils.uniform(-alpha_affine, alpha_affine, size=pts1.shape, random_state=random_state).astype(
-        np.float32
-    )
-    matrix = cv2.getAffineTransform(pts1, pts2)
-
-    warp_fn = _maybe_process_in_chunks(
-        cv2.warpAffine, M=matrix, dsize=(width, height), flags=interpolation, borderMode=border_mode, borderValue=value
-    )
-    img = warp_fn(img)
-
-    if approximate:
-        # Approximate computation smooth displacement map with a large enough kernel.
-        # On large images (512+) this is approximately 2X times faster
-        dx = random_utils.rand(height, width, random_state=random_state).astype(np.float32) * 2 - 1
-        cv2.GaussianBlur(dx, (17, 17), sigma, dst=dx)
-        dx *= alpha
-        if same_dxdy:
-            # Speed up even more
-            dy = dx
-        else:
-            dy = random_utils.rand(height, width, random_state=random_state).astype(np.float32) * 2 - 1
-            cv2.GaussianBlur(dy, (17, 17), sigma, dst=dy)
-            dy *= alpha
-    else:
-        dx = np.float32(
-            gaussian_filter((random_utils.rand(height, width, random_state=random_state) * 2 - 1), sigma) * alpha
-        )
-        if same_dxdy:
-            # Speed up
-            dy = dx
-        else:
-            dy = np.float32(
-                gaussian_filter((random_utils.rand(height, width, random_state=random_state) * 2 - 1), sigma) * alpha
-            )
-
-    x, y = np.meshgrid(np.arange(width), np.arange(height))
-
-    map_x = np.float32(x + dx)
-    map_y = np.float32(y + dy)
-
-    remap_fn = _maybe_process_in_chunks(
-        cv2.remap, map1=map_x, map2=map_y, interpolation=interpolation, borderMode=border_mode, borderValue=value
-    )
-    return remap_fn(img)
-
-
-
-from typing import Any, Dict, List, Optional, Sequence, Tuple, TypeVar, cast
-BoxInternalType = Tuple[float, float, float, float]
-BoxType = Union[BoxInternalType, Tuple[float, float, float, float, Any]]
-TBox = TypeVar("TBox", BoxType, BoxInternalType)
-
-
-def denormalize_bbox(bbox: TBox, rows: int, cols: int) -> TBox:
-    """Denormalize coordinates of a bounding box. Multiply x-coordinates by image width and y-coordinates
-    by image height. This is an inverse operation for :func:`~albumentations.augmentations.bbox.normalize_bbox`.
-
-    Args:
-        bbox: Normalized bounding box `(x_min, y_min, x_max, y_max)`.
-        rows: Image height.
-        cols: Image width.
-
-    Returns:
-        Denormalized bounding box `(x_min, y_min, x_max, y_max)`.
-
-    Raises:
-        ValueError: If rows or cols is less or equal zero
-
-    """
-    tail: Tuple[Any, ...]
-    (x_min, y_min, x_max, y_max), tail = bbox[:4], tuple(bbox[4:])
-
-    if rows <= 0:
-        raise ValueError("Argument rows must be positive integer")
-    if cols <= 0:
-        raise ValueError("Argument cols must be positive integer")
-
-    x_min, x_max = x_min * cols, x_max * cols
-    y_min, y_max = y_min * rows, y_max * rows
-
-    return cast(BoxType, (x_min, y_min, x_max, y_max) + tail)  # type: ignore
-
-
-def bbox_from_mask(mask):
-    """Create bounding box from binary mask (fast version)
-
-    Args:
-        mask (numpy.ndarray): binary mask.
-
-    Returns:
-        tuple: A bounding box tuple `(x_min, y_min, x_max, y_max)`.
-
-    """
-    rows = np.any(mask, axis=1)
-    if not rows.any():
-        return -1, -1, -1, -1
-    cols = np.any(mask, axis=0)
-    y_min, y_max = np.where(rows)[0][[0, -1]]
-    x_min, x_max = np.where(cols)[0][[0, -1]]
-    return x_min, y_min, x_max + 1, y_max + 1
-
-
-def normalize_bbox(bbox: TBox, rows: int, cols: int) -> TBox:
-    """Normalize coordinates of a bounding box. Divide x-coordinates by image width and y-coordinates
-    by image height.
-
-    Args:
-        bbox: Denormalized bounding box `(x_min, y_min, x_max, y_max)`.
-        rows: Image height.
-        cols: Image width.
-
-    Returns:
-        Normalized bounding box `(x_min, y_min, x_max, y_max)`.
-
-    Raises:
-        ValueError: If rows or cols is less or equal zero
-
-    """
-
-    if rows <= 0:
-        raise ValueError("Argument rows must be positive integer")
-    if cols <= 0:
-        raise ValueError("Argument cols must be positive integer")
-
-    tail: Tuple[Any, ...]
-    (x_min, y_min, x_max, y_max), tail = bbox[:4], tuple(bbox[4:])
-
-    x_min, x_max = x_min / cols, x_max / cols
-    y_min, y_max = y_min / rows, y_max / rows
-
-    return cast(BoxType, (x_min, y_min, x_max, y_max) + tail)  # type: ignore
+from augtools.img import functional as F
+import albumentations as A
 
 
 class ElasticTransform(DualTransform):
@@ -294,9 +71,10 @@ class ElasticTransform(DualTransform):
 
         res = {}
         random_state = random.randint(0, 10000)
+        h, w, _ = kwargs["img"].shape
 
         if "img" in kwargs:
-            res["img"] = elastic_transform(
+            res["img"] = F.elastic_transform(
                             img,
                             self.alpha,
                             self.sigma,
@@ -310,32 +88,21 @@ class ElasticTransform(DualTransform):
                             )
 
         if "bbox" in kwargs:
-            res["bbox"] = self.apply_to_bbox(kwargs["bbox"], random_state=random_state)
+            bboxes = []
+            for item in kwargs["bbox"]:
+                bboxes.append(self.apply_to_bbox(item, random_state=random_state, h=h, w=w))
+            res["bbox"] = bboxes
 
         return res
 
-    # def apply_to_mask(self, img, random_state=None, **params):
-    #     return F.elastic_transform(
-    #         img,
-    #         self.alpha,
-    #         self.sigma,
-    #         self.alpha_affine,
-    #         cv2.INTER_NEAREST,
-    #         self.border_mode,
-    #         self.mask_value,
-    #         np.random.RandomState(random_state),
-    #         self.approximate,
-    #         self.same_dxdy,
-    #     )
-
-    def apply_to_bbox(self, bbox, random_state=None, **params):
-        rows, cols = params["rows"], params["cols"]
+    def apply_to_bbox(self, bbox, random_state=None, h=None, w=None):
+        rows, cols = h, w
         mask = np.zeros((rows, cols), dtype=np.uint8)
-        bbox_denorm = denormalize_bbox(bbox, rows, cols)
+        bbox_denorm = F.denormalize_bbox(bbox, rows, cols)
         x_min, y_min, x_max, y_max = bbox_denorm[:4]
         x_min, y_min, x_max, y_max = int(x_min), int(y_min), int(x_max), int(y_max)
         mask[y_min:y_max, x_min:x_max] = 1
-        mask = elastic_transform(
+        mask = F.elastic_transform(
             mask,
             self.alpha,
             self.sigma,
@@ -346,10 +113,9 @@ class ElasticTransform(DualTransform):
             np.random.RandomState(random_state),
             self.approximate,
         )
-        bbox_returned = bbox_from_mask(mask)
-        bbox_returned = normalize_bbox(bbox_returned, rows, cols)
+        bbox_returned = F.bbox_from_mask(mask)
+        bbox_returned = F.normalize_bbox(bbox_returned, rows, cols)
         return bbox_returned
-
 
 
 if __name__ == '__main__':
@@ -359,11 +125,16 @@ if __name__ == '__main__':
     image = prefix + 'test.jpg'
 
     img = read_image(image)
-    # print(img)
-    bbox = (50, 60, 50, 80)
+    print(img.shape)
+    bbox = [(170 / 500, 30 / 375, 300 / 500, 220 / 375)]
+    keypoint = [(230, 80, 1, 1)]
+
+    show_bbox_keypoint_image_float(img, bbox=bbox, keypoint=keypoint)
 
     transform = ElasticTransform()
-    result = transform(img=img, force_apply=True)
-    # print(result['img'])
-    # print(result["bbox"])
-    show_image(result['img'])
+    re = transform(img=img, force_apply=True, bbox=bbox, keypoint=keypoint)
+    show_bbox_keypoint_image_float(re['img'], bbox=re['bbox'], keypoint=None)
+
+    cc = A.ElasticTransform(always_apply=True)
+    result = cc(image=img, bboxes=bbox)
+    show_bbox_keypoint_image_float(result['image'], bbox=result['bboxes'], keypoint=None)
