+++ augtools/img/transforms/geometric/HorizontalFlip.py
@@ -1,12 +1,7 @@
-import warnings
-import random
-from itertools import product
-
-import cv2
-import numpy as np
-
 from augtools.img.transform import DualTransform
-import functional as F
+import albumentations as A
+import random
+import augtools.img.functional as F
 
 
 class HorizontalFlip(DualTransform):
@@ -22,16 +17,28 @@ class HorizontalFlip(DualTransform):
         uint8, float32
     """
 
-    def _compute_x_function(self, img, rs=None):
-        return F.hflip(img)
+    def __call__(self, *args, force_apply: bool = False, **kwargs):
+        if (random.random() < self.p) and not self.always_apply and not force_apply:
+            return kwargs
+        res = {}
+        h, w, _ = kwargs["img"].shape
 
-    def _compute_bbox_function(self, y, rs=None):
-        return F.bbox_hflip(y)
+        if "img" in kwargs:
+            res["img"] = F.hflip(img)
 
-    # def _compute_keypoint_function(self, y, rs=None):
-    #    return F.keypoint_hflip(keypoint, **params)
+        if "bbox" in kwargs:
+            bboxes = []
+            for item in kwargs["bbox"]:
+                bboxes.append(F.bbox_hflip(item, h, w))
+            res["bbox"] = bboxes
 
+        if "keypoint" in kwargs:
+            points = []
+            for item in kwargs["keypoint"]:
+                points.append(F.keypoint_hflip(item, h, w))
+            res["keypoint"] = points
 
+        return res
 
 
 if __name__ == '__main__':
@@ -41,11 +48,16 @@ if __name__ == '__main__':
     image = prefix + 'test.jpg'
 
     img = read_image(image)
-    # bboxs = [(50, 60, 50, 80), (50, 60, 50, 80)]
-    bbox = (50, 60, 50, 80)
+    print(img.shape)
+    bbox = [(170, 30, 300, 220)]
+    keypoint = [(230, 80, 1, 1)]
+
+    show_bbox_keypoint_image(img, bbox=bbox, keypoint=keypoint)
 
     transform = HorizontalFlip()
-    result = transform(img=img, bbox=bbox, force_apply=True)
-    print(result['img'])
+    re = transform(img=img, force_apply=True, bbox=bbox, keypoint=keypoint)
+    show_bbox_keypoint_image(re['img'], bbox=re['bbox'], keypoint=re['keypoint'])
 
-    show_image(result['img'])
+    cc = A.HorizontalFlip(always_apply=True)
+    result = cc(image=img, bboxes=bbox, keypoints=keypoint)
+    show_bbox_keypoint_image(result['image'], bbox=result['bboxes'], keypoint=result['keypoints'])
