+++ augtools/img/transforms/geometric/Transpose.py
@@ -1,73 +1,7 @@
 from augtools.img.transform import DualTransform
-from augtools.img.transforms.utils.img_utils import *
-import skimage as sk
 import random
-from typing_extensions import Concatenate, ParamSpec
-import math
-
-NumType = Union[int, float, np.ndarray]
-BoxInternalType = Tuple[float, float, float, float]
-BoxType = Union[BoxInternalType, Tuple[float, float, float, float, Any]]
-KeypointInternalType = Tuple[float, float, float, float]
-KeypointType = Union[KeypointInternalType, Tuple[float, float, float, float, Any]]
-ImageColorType = Union[float, Sequence[float]]
-
-ScaleFloatType = Union[float, Tuple[float, float]]
-ScaleIntType = Union[int, Tuple[int, int]]
-
-FillValueType = Optional[Union[int, float, Sequence[int], Sequence[float]]]
-
-
-def transpose(img: np.ndarray) -> np.ndarray:
-    return img.transpose(1, 0, 2) if len(img.shape) > 2 else img.transpose(1, 0)
-
-
-def bbox_transpose(
-    bbox: KeypointInternalType, axis: int
-) -> KeypointInternalType:  # skipcq: PYL-W0613
-    """Transposes a bounding box along given axis.
-
-    Args:
-        bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
-        axis: 0 - main axis, 1 - secondary axis.
-        rows: Image rows.
-        cols: Image cols.
-
-    Returns:
-        A bounding box tuple `(x_min, y_min, x_max, y_max)`.
-
-    Raises:
-        ValueError: If axis not equal to 0 or 1.
-
-    """
-    x_min, y_min, x_max, y_max = bbox[:4]
-    if axis not in {0, 1}:
-        raise ValueError("Axis must be either 0 or 1.")
-    if axis == 0:
-        bbox = (y_min, x_min, y_max, x_max)
-    if axis == 1:
-        bbox = (1 - y_max, 1 - x_max, 1 - y_min, 1 - x_min)
-    return bbox
-
-
-def keypoint_transpose(keypoint: KeypointInternalType) -> KeypointInternalType:
-    """Rotate a keypoint by angle.
-
-    Args:
-        keypoint: A keypoint `(x, y, angle, scale)`.
-
-    Returns:
-        A keypoint `(x, y, angle, scale)`.
-
-    """
-    x, y, angle, scale = keypoint[:4]
-
-    if angle <= np.pi:
-        angle = np.pi - angle
-    else:
-        angle = 3 * np.pi - angle
-
-    return y, x, angle, scale
+import augtools.img.functional as F
+import albumentations as A
 
 
 class Transpose(DualTransform):
@@ -89,13 +23,19 @@ class Transpose(DualTransform):
         res = {}
 
         if "img" in kwargs:
-            res["img"] = transpose(kwargs["img"])
+            res["img"] = F.transpose(kwargs["img"])
 
         if "bbox" in kwargs:
-            res["bbox"] = bbox_transpose(kwargs["bbox"], 0)
+            bboxes = []
+            for item in kwargs["bbox"]:
+                bboxes.append(F.bbox_transpose(item, 0))
+            res["bbox"] = bboxes
 
         if "keypoint" in kwargs:
-            res["keypoint"] = keypoint_transpose(kwargs["keypoint"])
+            points = []
+            for item in kwargs["keypoint"]:
+                points.append(F.keypoint_transpose(item))
+            res["keypoint"] = points
 
         return res
 
@@ -107,11 +47,16 @@ if __name__ == '__main__':
     image = prefix + 'test.jpg'
 
     img = read_image(image)
-    # print(img)
-    bbox = (50, 60, 50, 80)
+    print(img.shape)
+    bbox = [(170 / 500, 30 / 375, 300 / 500, 220 / 375)]
+    keypoint = [(230, 80, 1, 1)]
+
+    show_bbox_keypoint_image_float(img, bbox=bbox, keypoint=keypoint)
 
     transform = Transpose()
-    result = transform(img=img, force_apply=True, bbox=bbox)
-    # print(result['img'])
-    print(result["bbox"])
-    show_image(result['img'])
+    re = transform(img=img, force_apply=True, bbox=bbox, keypoint=keypoint)
+    show_bbox_keypoint_image_float(re['img'], bbox=re['bbox'], keypoint=re['keypoint'])
+
+    cc = A.Transpose(always_apply=True)
+    result = cc(image=img, bboxes=bbox, keypoints=keypoint)
+    show_bbox_keypoint_image_float(result['image'], bbox=result['bboxes'], keypoint=result['keypoints'])
