+++ augtools/img/transforms/geometric/VerticalFlip.py
@@ -1,16 +1,8 @@
-import warnings
-import random
-from itertools import product
-
-import cv2
-import numpy as np
-
 from augtools.img.transform import DualTransform
-import functional as F
+import augtools.img.functional as F
+import random
+import albumentations as A
 
-# DualTransform 是对 y 也需要处理
-# BasicTrans **kwargs 传入，然后 BasicTrans _compute_x/y 传入是 **kwargs
-#
 
 class VerticalFlip(DualTransform):
     """Flip the input vertically around the x-axis.
@@ -25,14 +17,28 @@ class VerticalFlip(DualTransform):
         uint8, float32
     """
 
-    def _compute_x_function(self, img, rs=None):
-        return F.vflip(img)
+    def __call__(self, *args, force_apply: bool = False, **kwargs):
+        if (random.random() < self.p) and not self.always_apply and not force_apply:
+            return kwargs
+        res = {}
+        h, w, _ = kwargs["img"].shape
 
-    def _compute_bbox_function(self, y, rs=None):
-        return F.bbox_vflip(y)
+        if "img" in kwargs:
+            res["img"] = F.vflip(img)
 
-    # def _compute_keypoint_function(self, y, rs=None):
-    #     return F.keypoint_vflip(y, **params)
+        if "bbox" in kwargs:
+            bboxes = []
+            for item in kwargs["bbox"]:
+                bboxes.append(F.bbox_vflip(item, h, w))
+            res["bbox"] = bboxes
+
+        if "keypoint" in kwargs:
+            points = []
+            for item in kwargs["keypoint"]:
+                points.append(F.keypoint_vflip(item, h, w))
+            res["keypoint"] = points
+
+        return res
 
 
 if __name__ == '__main__':
@@ -42,12 +48,16 @@ if __name__ == '__main__':
     image = prefix + 'test.jpg'
 
     img = read_image(image)
-    # bboxs = [(50, 60, 50, 80), (50, 60, 50, 80)]
-    bbox = (50, 60, 50, 80)
+    print(img.shape)
+    bbox = [(170/500, 30/375, 300/500, 220/375)]
+    keypoint = [(230, 80, 1, 1)]
 
-    transform = VerticalFlip()
-    result = transform(img=img, bbox=bbox, force_apply=True)
-    print(result['img'])
-    print(result['bbox'])
+    show_bbox_keypoint_image_float(img, bbox=bbox, keypoint=keypoint)
 
-    show_image(result['img'])
+    # transform = VerticalFlip()
+    # re = transform(img=img, force_apply=True, bbox=bbox, keypoint=keypoint)
+    # show_bbox_keypoint_image(re['img'], bbox=re['bbox'], keypoint=re['keypoint'])
+
+    cc = A.VerticalFlip(always_apply=True)
+    result = cc(image=img, bboxes=bbox, keypoints=keypoint)
+    show_bbox_keypoint_image_float(result['image'], bbox=result['bboxes'], keypoint=result['keypoints'])
