--- augtools/img/transforms/geometric/functional.py
@@ -1,97 +1,0 @@
-import numpy as np
-import cv2
-
-
-def vflip(img: np.ndarray) -> np.ndarray:
-    return np.ascontiguousarray(img[::-1, ...])
-
-
-def hflip(img: np.ndarray) -> np.ndarray:
-    return np.ascontiguousarray(img[:, ::-1, ...])
-
-
-def hflip_cv2(img: np.ndarray) -> np.ndarray:
-    return cv2.flip(img, 1)
-
-
-def random_flip(img: np.ndarray, code: int) -> np.ndarray:
-    return cv2.flip(img, code)
-
-
-# Tuple[float, float, float, float]
-def bbox_vflip(bbox):
-    """Flip a bounding box vertically around the x-axis.
-
-    Args:
-        bbox(Tuple[float, float, float, float]): A bounding box `(x_min, y_min, x_max, y_max)`.
-        rows: Image rows.
-        cols: Image cols.
-
-    Returns:
-        tuple: A bounding box `(x_min, y_min, x_max, y_max)`.
-
-    """
-    x_min, y_min, x_max, y_max = bbox[:4]
-    return (x_min, 1 - y_max, x_max, 1 - y_min)
-
-
-def bbox_hflip(bbox):
-    """Flip a bounding box horizontally around the y-axis.
-
-    Args:
-        bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
-        rows: Image rows.
-        cols: Image cols.
-
-    Returns:
-        A bounding box `(x_min, y_min, x_max, y_max)`.
-
-    """
-    x_min, y_min, x_max, y_max = bbox[:4]
-    return 1 - x_max, y_min, 1 - x_min, y_max
-
-
-def bbox_flip(bbox, d: int, rows: int, cols: int):
-    """Flip a bounding box either vertically, horizontally or both depending on the value of `d`.
-
-    Args:
-        bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
-        d: dimension. 0 for vertical flip, 1 for horizontal, -1 for transpose
-        rows: Image rows.
-        cols: Image cols.
-
-    Returns:
-        A bounding box `(x_min, y_min, x_max, y_max)`.
-
-    Raises:
-        ValueError: if value of `d` is not -1, 0 or 1.
-
-    """
-    if d == 0:
-        bbox = bbox_vflip(bbox, rows, cols)
-    elif d == 1:
-        bbox = bbox_hflip(bbox, rows, cols)
-    elif d == -1:
-        bbox = bbox_hflip(bbox, rows, cols)
-        bbox = bbox_vflip(bbox, rows, cols)
-    else:
-        raise ValueError("Invalid d value {}. Valid values are -1, 0 and 1".format(d))
-    return bbox
-
-
-# Tuple[float, float, float, float]
-def keypoint_vflip(keypoint, rows: int, cols: int):
-    """Flip a keypoint vertically around the x-axis.
-
-    Args:
-        keypoint(Tuple[float, float, float, float]): A keypoint `(x, y, angle, scale)`.
-        rows: Image height.
-        cols: Image width.
-
-    Returns:
-        tuple: A keypoint `(x, y, angle, scale)`.
-
-    """
-    x, y, angle, scale = keypoint[:4]
-    angle = -angle
-    return (x, (rows - 1) - y, angle, scale)
